pyproject.toml
[project]
name = "lce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pypdf>=6.9.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "tzdata>=2025.3",
    "uvicorn[standard]>=0.43.0",
]

[dependency-groups]
dev = [
    "filelock>=3.25.2",
    "psutil>=7.2.2",
    "pytest-xdist>=3.8.0",
    "pytest~=7.4.4",
    "pytest-html>=4.0.0",
    "pytest-html-reporter>=0.2.6",
    "langchain-chroma>=1.1.0",
    "sentence-transformers>=5.3.0",
]

#-----------------------------------------------------------------------------
# Build system
# Use "setuptools.build_meta" (classic backend) — works with all setuptools
# versions. The "setuptools.backends.legacy:build" form requires setuptools 69+
# and is NOT needed here.
# -----------------------------------------------------------------------------
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

# -----------------------------------------------------------------------------
# Pytest — single source of truth (pytest.ini deleted)
# -----------------------------------------------------------------------------
[tool.pytest.ini_options]
pythonpath = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "llm_ollama_models: a marker to run llm_ollama_models",
    "llm_open_ai_models: a marker to run llm_open_ai_models",
    "llm_platform_provider_open_ai: a marker to run llm_platform_provider_open_ai"
]
